Add BasePageBlock and move TestBlock to test_models.py
I moved TestBlock outside models.py mainly so I could inherit BasePageBlock
from it, so that code gets tested. I think it makes more sense to have the
TestBlock in a different file than the actual pagetree models anyways.
