+++ pagetree/__init__.py
@@ -1,0 +1,3 @@
+# This is required for south's syncdb to pick up TestBlock in
+# Django 1.6
+from pagetree.test_models import *
